Move key events into fixtures
